requirements.txt
fastapi~=0.115.13
uvicorn[standard]
sqlalchemy~=2.0.41
psycopg2-binary
python-dotenv
ccxt~=4.4.90
pydantic-settings
pytest~=8.4.1
pytest-asyncio
pandas~=2.3.0
pandas-ta
numpy~=2.3.0
scipy
pandas_ta~=0.3.14b0
pydantic~=2.11.7
mexc-api
aiohttp
